ci(rust): actually run clippy, and clean the workspace to zero warnings
The rust-lint job has installed clippy since it was written
(`components: rustfmt, clippy`) but its only step was
`cargo fmt --all -- --check`. Clippy has never run in CI, so 81 warning
sites accumulated across the workspace unnoticed.

This clears all of them and gates the job at `-D warnings`.

Most were mechanical and applied with `cargo clippy --fix`: 30 redundant
closures in runtime_config.rs, 8 manual div_ceil, 7 no-op casts, 7
needless borrows, 4 let-and-return, 3 derivable Default impls, and a
dozen singletons. Fixed by hand: a `.clone()` on the Copy type
`Statistics`, a missing `Default for Executor`, an index loop in a
vector_machine test rewritten as a slice fill, and formatting `--fix`
left for rustfmt.

`SimpleTiming::new` triggered too_many_arguments. Rather than suppress
it, the function is gone: it was `pub` with no caller anywhere in
PLENA_Simulator, PLENA_Compiler or PLENA_RTL except the adjacent
`preset_ddr4_2400p`, which passed eight positional literals
(`833ps, 16, 16, 16, 64, 16, 4`) that were unreadable at the call site.
Its body is now inlined into that sole caller with the constants named,
which removes the lint, an unused public API, and the positional
literals at once.

Eight `#[allow]` attributes remain, covering four lints, each with a
short note on why the lint is a false positive there rather than muting
it globally:

  - executor.rs `enter()` — await_holding_lock: the guard is dropped
    before the await and re-acquired after; clippy flags it only because
    the binding is reassigned inside the loop.
  - frfcfs.rs `acquire()` — deref_addrof: clippy's suggestion does not
    compile (E0597/E0505/E0716); the raw round-trip launders the borrow
    to 'static, which `Inner::req` requires.
  - ddr4.rs / hbm2.rs `resolve_n_*` (5 fns, 7 warning sites — the two
    with a per-dq inner match fire twice) — match_overlapping_arm: the
    `..=N` arms are an ordered stair-step over JEDEC speed bins, and
    rewriting ~30 bounds as disjoint ranges risks a silent transcription
    error that no test would catch.
  - stage_profile.rs `classify_comment()` — if_same_then_else: two
    adjacent rules match different comments but yield the same StageKind.

Behavior unchanged, verified on this merge base before and after:

  cargo test --workspace   178 passed, 0 failed  ->  178 passed, 0 failed
  cargo fmt --all --check  pass                  ->  pass
  cargo clippy             81 warning sites      ->  0
  cargo build --release    ok                    ->  ok

End-to-end, testbench/direct_emit/silu_test.py run against the rebuilt
emulator produces byte-identical artifacts (sha256 over
generated_asm_code.asm, comparison_params.json, fp_sram.bin,
int_sram.bin, hbm_for_behave_sim.bin, act_tensor.pt — all 6 unchanged).

The toolchain is not on PATH by default, which is why this went
unnoticed: `nix develop` supplies cargo/clippy plus Ramulator2 and
libtorch, and the Python side needs the uv venv interpreter rather than
the devShell's python.

+++ transactional_emulator/src/op.rs
@@ -347,8 +347,8 @@ impl Opcode {
         let rs2 = ((instr >> (OPCODE_WIDTH + OPERAND_WIDTH * 2)) & mask(OPERAND_WIDTH)) as u8;
         let rs3 = ((instr >> (OPCODE_WIDTH + OPERAND_WIDTH * 3)) & mask(OPERAND_WIDTH)) as u8;
         let funct1 = ((instr >> (OPCODE_WIDTH + OPERAND_WIDTH * 4)) & mask(OPERAND_WIDTH)) as u8;
-        let imm = ((instr >> (OPCODE_WIDTH + OPERAND_WIDTH)) & mask(IMM_WIDTH)) as u32;
-        let imm2 = ((instr >> (OPCODE_WIDTH + OPERAND_WIDTH * 2)) & mask(IMM_2_WIDTH)) as u32;
+        let imm = (instr >> (OPCODE_WIDTH + OPERAND_WIDTH)) & mask(IMM_WIDTH);
+        let imm2 = (instr >> (OPCODE_WIDTH + OPERAND_WIDTH * 2)) & mask(IMM_2_WIDTH);
 
         match opcode {
             0x00 => Self::Invalid,
